requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.3.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.1
pydantic_core==2.41.3
pymongo==4.15.3
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
